[IMP] base: promote slugify and slugify_one to base
Pull #168691 made the various slugify function composable and added a
base implementation of slug and unslug. It turns out a base
implementation of slugify and slugify_one are also necessary.
